MGS-3509 use test when building on jenkins
rearrange to allow same pytest.ini for setup.py (jenkins) and Makefile (dev).
remove RuntimeError if envs not set for test purposes.

+++ cli.py
@@ -21,12 +21,6 @@ args = None
     help="Stop for debugging.",
     is_flag=True,
 )
-# @click.option(
-#     '--verbose',
-#     default=False,
-#     help="Output large quantity of information about what the script is doing.",
-#     is_flag=True,
-# )
 @click.command()
 def main(dry_run, debug):
     """Console script for repo_clean."""
@@ -35,6 +29,10 @@ def main(dry_run, debug):
         set_trace()
 
     repo = Repo()
+    if not repo.nexus.host:
+        print("Invalid config")
+        exit(-1)
+
     purge_list = repo.purge_list()
     if dry_run or purge_list is None:
         return
